ai_service/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-pinecone
langchain-text-splitters
pinecone-client
openai
unstructured[pdf]
pypdf
tiktoken
python-multipart
python-dotenv
pydantic-settings
pytest
httpx